// src/lib/models/BusinessActor.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BusinessActor = {
    id?: string;
    name?: string;
    businessId?: string;
    niu?: string;
    tradeRegistryNumber?: string;
    website?: string;
    contactPhone?: string;
    privateAddress?: string;
    businessAddress?: string;
    businessProfile?: string;
    createdAt?: string;
};