// src/lib/models/OrganizationMember.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrganizationMember = {
    id?: string;
    organizationId?: string;
    userId?: string;
    userEmail?: string;
    userFirstName?: string;
    userLastName?: string;
    agencyId?: string;
    agencyName?: string;
    roleId?: string;
    roleName?: string;
    joinedAt?: string;
    active?: boolean;
};